ERP-26 :: Resolve all PR Comments also do some Additional code related changes

+++ MidCapERP.Dto/Product/ProductMainRequestDto.cs
@@ -14,7 +14,7 @@ namespace MidCapERP.Dto.Product
             ProductMaterialRequestDto = new List<ProductMaterialRequestDto>();
         }
 
-        public int ProductId { get; set; } = 0;
+        public Int64 ProductId { get; set; } = 0;
         [DisplayName("Cost Price")]
         public decimal CostPrice { get; set; }
         [DisplayName("Retailer Price")]
